pyproject.toml
[project]
name = "makerai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "openai-agents>=0.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
